# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Edubee Camp API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /leads:
    get:
      operationId: listLeads
      tags: [applications]
      summary: List all leads
      security:
        - BearerAuth: []
      parameters:
        - name: agentId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadListResponse"
    post:
      operationId: createLead
      tags: [applications]
      summary: Create a new lead
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadRequest"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [applications]
      summary: Get lead by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    put:
      operationId: updateLead
      tags: [applications]
      summary: Update lead
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadRequest"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [applications]
      summary: Delete lead
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List all applications
      security:
        - BearerAuth: []
      parameters:
        - name: agentId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: packageGroupId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationListResponse"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create a new application
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationRequest"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationDetail"
    put:
      operationId: updateApplication
      tags: [applications]
      summary: Update application
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationRequest"
      responses:
        "200":
          description: Application updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/status:
    put:
      operationId: updateApplicationStatus
      tags: [applications]
      summary: Update application status
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - total
        - page
        - limit
        - totalPages
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - email
        - password
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - accessToken
        - refreshToken
        - user
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - super_admin
            - admin
            - camp_coordinator
            - education_agent
            - partner_institute
            - partner_hotel
            - partner_pickup
            - partner_tour
            - parent_client
        fullName:
          type: string
        phone:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        lineId:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        preferredLang:
          type: string
          nullable: true
        companyName:
          type: string
          nullable: true
        businessRegNo:
          type: string
          nullable: true
        countryOfOps:
          type: string
          nullable: true
        platformCommRate:
          type: string
          nullable: true
        status:
          type: string
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - fullName
        - status
        - createdAt
        - updatedAt
...
    Lead:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
          nullable: true
        fullName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        nationality:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        interestedIn:
          type: string
          nullable: true
        status:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - status
        - createdAt
        - updatedAt
    CreateLeadRequest:
      type: object
      properties:
        agentId:
          type: string
          nullable: true
        fullName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        nationality:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        interestedIn:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - fullName
    LeadListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    Application:
      type: object
      properties:
        id:
          type: string
        applicationNumber:
          type: string
          nullable: true
        agentId:
          type: string
          nullable: true
        clientId:
          type: string
          nullable: true
        packageGroupId:
          type: string
          nullable: true
        packageId:
          type: string
          nullable: true
        preferredStartDate:
          type: string
          format: date
          nullable: true
        status:
          type: string
        totalChildren:
          type: integer
        totalAdults:
          type: integer
        primaryLanguage:
          type: string
        referralSource:
          type: string
          nullable: true
        specialRequests:
          type: string
          nullable: true
        termsAccepted:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - totalChildren
        - totalAdults
        - primaryLanguage
        - termsAccepted
        - createdAt
        - updatedAt
    ApplicationParticipant:
      type: object
      properties:
        id:
          type: string
        applicationId:
          type: string
          nullable: true
        participantType:
          type: string
        sequenceOrder:
          type: integer
        fullName:
          type: string
        dateOfBirth:
          type: string
          format: date
          nullable: true
        gender:
          type: string
          nullable: true
        nationality:
          type: string
          nullable: true
        passportNumber:
          type: string
          nullable: true
        passportExpiry:
          type: string
          format: date
          nullable: true
        grade:
          type: string
          nullable: true
        schoolName:
          type: string
          nullable: true
        englishLevel:
          type: string
          nullable: true
        medicalConditions:
          type: string
          nullable: true
        dietaryRequirements:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - participantType
        - sequenceOrder
        - fullName
        - createdAt
    ApplicationDetail:
      allOf:
        - $ref: "#/components/schemas/Application"
        - type: object
          properties:
            participants:
              type: array
              items:
                $ref: "#/components/schemas/ApplicationParticipant"
    CreateApplicationRequest:
      type: object
      properties:
        agentId:
          type: string
          nullable: true
        clientId:
          type: string
          nullable: true
        packageGroupId:
          type: string
          nullable: true
        packageId:
          type: string
          nullable: true
        preferredStartDate:
          type: string
          format: date
          nullable: true
        status:
          type: string
          nullable: true
        totalChildren:
          type: integer
          nullable: true
        totalAdults:
          type: integer
          nullable: true
        primaryLanguage:
          type: string
          nullable: true
        referralSource:
          type: string
          nullable: true
        specialRequests:
          type: string
          nullable: true
        termsAccepted:
          type: boolean
          nullable: true
        notes:
          type: string
          nullable: true
    UpdateApplicationRequest:
      type: object
      properties:
        agentId:
          type: string
          nullable: true
        clientId:
          type: string
          nullable: true
        packageGroupId:
          type: string
          nullable: true
        packageId:
          type: string
          nullable: true
        preferredStartDate:
          type: string
          format: date
          nullable: true
        totalChildren:
          type: integer
          nullable: true
        totalAdults:
          type: integer
          nullable: true
        primaryLanguage:
          type: string
          nullable: true
        referralSource:
          type: string
          nullable: true
        specialRequests:
          type: string
          nullable: true
        termsAccepted:
          type: boolean
          nullable: true
        notes:
          type: string
          nullable: true
    UpdateStatusRequest:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - status
    ApplicationListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
...