# lib/api-zod/src/generated/types/applicationDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Edubee Camp API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from "./application";
import type { ApplicationParticipant } from "./applicationParticipant";

export type ApplicationDetail = Application & {
  participants?: ApplicationParticipant[];
};