# lib/api-zod/src/generated/types/applicationListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Edubee Camp API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from "./application";
import type { PaginationMeta } from "./paginationMeta";

export interface ApplicationListResponse {
  data: Application[];
  meta: PaginationMeta;
}