# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Edubee Camp API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: string;
  email: string;
  role: UserRole;
  fullName: string;
  phone?: string | null;
  whatsapp?: string | null;
  lineId?: string | null;
  avatarUrl?: string | null;
  timezone?: string | null;
  preferredLang?: string | null;
  companyName?: string | null;
  businessRegNo?: string | null;
  countryOfOps?: string | null;
  platformCommRate?: string | null;
  status: string;
  lastLoginAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}